Made changes in the Fetchdata
