+++ package.json
@@ -60,6 +60,9 @@
     "react-axe": "^3.4.1",
     "react-docgen-typescript-loader": "^3.2.1",
     "react-dom": "^16.11.0",
+    "react-popper": "2.2.3",
+    "reactstrap": "8.4.1",
+    "react-transition-group": "4.4.1",
     "regenerator-runtime": "^0.13.3",
     "rimraf": "^3.0.2",
     "style-loader": "^1.2.1",
